+++ azure-batch/azure/batch/models/network_configuration.py
@@ -30,15 +30,16 @@ class NetworkConfiguration(Model):
      checking if the specified VNet has any associated Network Security Groups
      (NSG). If communication to the compute nodes in the specified subnet is
      denied by an NSG, then the Batch service will set the state of the compute
-     nodes to unusable. For pools created via virtualMachineConfiguration the
-     Batch account must have poolAllocationMode userSubscription in order to
-     use a VNet. If the specified VNet has any associated Network Security
-     Groups (NSG), then a few reserved system ports must be enabled for inbound
-     communication. For pools created with a virtual machine configuration,
-     enable ports 29876 and 29877, as well as port 22 for Linux and port 3389
-     for Windows. For pools created with a cloud service configuration, enable
-     ports 10100, 20100, and 30100. Also enable outbound connections to Azure
-     Storage on port 443. For more details see:
+     nodes to unusable. For pools created with virtualMachineConfiguration only
+     ARM virtual networks ('Microsoft.Network/virtualNetworks') are supported,
+     but for pools created with cloudServiceConfiguration both ARM and classic
+     virtual networks are supported. If the specified VNet has any associated
+     Network Security Groups (NSG), then a few reserved system ports must be
+     enabled for inbound communication. For pools created with a virtual
+     machine configuration, enable ports 29876 and 29877, as well as port 22
+     for Linux and port 3389 for Windows. For pools created with a cloud
+     service configuration, enable ports 10100, 20100, and 30100. Also enable
+     outbound connections to Azure Storage on port 443. For more details see:
      https://docs.microsoft.com/en-us/azure/batch/batch-api-basics#virtual-network-vnet-and-firewall-configuration
     :type subnet_id: str
     :param endpoint_configuration: The configuration for endpoints on compute