+++ azure-batch/azure/batch/operations/task_operations.py
@@ -22,7 +22,7 @@ class TaskOperations(object):
     :param config: Configuration of service client.
     :param serializer: An object model serializer.
     :param deserializer: An objec model deserializer.
-    :ivar api_version: Client API Version. Constant value: "2017-06-01.5.1".
+    :ivar api_version: Client API Version. Constant value: "2017-09-01.6.0".
     """
 
     def __init__(self, client, config, serializer, deserializer):
@@ -30,13 +30,18 @@ class TaskOperations(object):
         self._client = client
         self._serialize = serializer
         self._deserialize = deserializer
-        self.api_version = "2017-06-01.5.1"
+        self.api_version = "2017-09-01.6.0"
 
         self.config = config
 
     def add(
             self, job_id, task, task_add_options=None, custom_headers=None, raw=False, **operation_config):
         """Adds a task to the specified job.
+
+        The maximum lifetime of a task from addition to completion is 7 days.
+        If a task has not completed within 7 days of being added it will be
+        terminated by the Batch service and left in whatever state it was in at
+        that time.
 
         :param job_id: The ID of the job to which the task is to be added.
         :type job_id: str
@@ -251,7 +256,11 @@ class TaskOperations(object):
         will not create extra tasks unexpectedly. If the response contains any
         tasks which failed to add, a client can retry the request. In a retry,
         it is most efficient to resubmit only tasks that failed to add, and to
-        omit tasks that were successfully added on the first attempt.
+        omit tasks that were successfully added on the first attempt. The
+        maximum lifetime of a task from addition to completion is 7 days. If a
+        task has not completed within 7 days of being added it will be
+        terminated by the Batch service and left in whatever state it was in at
+        that time.
 
         :param job_id: The ID of the job to which the task collection is to be
          added.